+++ packages/api-lib/tests/integration/ingestCollections.js
@@ -3,9 +3,7 @@ const ingest = require('../../libs/ingest').ingest
 const backend = require('../../libs/es')
 
 async function doIngest() {
-  //ingest('https://landsat-stac.s3.amazonaws.com/landsat-8-l1/catalog.json', backend)
-  await ingest('../fixtures/stac/catalog.json', backend)
-  console.log('Items done')
+  await ingest('../fixtures/stac/catalog.json', backend, true, true)
+  console.log('Collections done')
 }
 doIngest()
-