+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.2.3] - 2019-01-29
+
+### Fixed
+- Proper handling of bounding box passed as string
+
+### Changed
+- De-normalize Item properties to include all properties from collection
+- Flattened elastic search to simplify query logic
+- Items returned will now include all 'Common' properties that are in the Items Collection
+
 ## [v0.2.2] - 2019-01-21
 
 ### Fixed
@@ -74,7 +84,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Refactor and improve splitting
 
 [Unreleased]: https://github.com/sat-utils/sat-api/compare/master...develop
-[v0.2.2]: https://github.com/sat-utils/sat-api/compare/v0.2.0...v0.2.2
+[v0.2.3]: https://github.com/sat-utils/sat-api/compare/v0.2.2...v0.2.3
+[v0.2.2]: https://github.com/sat-utils/sat-api/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/sat-utils/sat-api/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/sat-utils/sat-api/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/sat-utils/sat-api/compare/v0.0.2...v0.1.0